Add Item class
Adds an Item class to represent items in game. Takes in the item_id (like minecraft:diamond_sword) and a quantity (defaults to 1). It also supports damage for items. I wanted to add enchantment support but I didn't know the best way to do that. I was thinking I could make an enchantment class but then it could be a top-level class or an inner class. I feel an inner class would be better but I'm not sure. Another option is a tuple that just contains the enchantment and the level. Since I couldn't decide, I didn't make it. Perhaps you could give advice on how to handle enchantments. I tested it and it works perfectly.